+++ Libft/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: obelaizi <obelaizi@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/08 14:06:11 by obelaizi          #+#    #+#             */
-/*   Updated: 2022/10/12 00:51:05 by obelaizi         ###   ########.fr       */
+/*   Updated: 2022/10/31 21:43:41 by obelaizi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (s[count])
